requirements.txt
python-docx==1.1.0
qdrant-client==1.8.0
langchain-openai==0.0.8
pathlib==1.0.1
django==5.0.3
waitress==3.0.0